Trait pallet_remote_treasury::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>;
    type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize;
    type PalletId: Get<PalletId>;
    type SelfAssetId: Get<Self::AssetId>;
    type RelayChainAssetId: Get<Self::AssetId>;
    type XcmAssetTransfer: XcmTransfer<Self::AccountId, Self::Balance, Self::AssetId>;
    type AssetIdConvert: Convert<Self::AssetId, Option<MultiLocation>>;
    type AccountId32Convert: Convert<Self::AccountId, [u8; 32]>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type AdminOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to manage the treasury balance and initiate XCM withdrawals

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>

The balance type for cross chain transfers

type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize

Asset Id that is used to identify different kinds of assets.

type PalletId: Get<PalletId>

PalletId used to generate the AccountId which holds the balance of the treasury.

type SelfAssetId: Get<Self::AssetId>

The native asset id

type RelayChainAssetId: Get<Self::AssetId>

Identifier for the relay chain’s specific asset

type XcmAssetTransfer: XcmTransfer<Self::AccountId, Self::Balance, Self::AssetId>

The interface to Cross-chain transfer.

type AssetIdConvert: Convert<Self::AssetId, Option<MultiLocation>>

Convert a T::AssetId to its relative MultiLocation identifier.

type AccountId32Convert: Convert<Self::AccountId, [u8; 32]>

Convert Self::Account to AccountId32

type WeightInfo: WeightInfo

The weight for this pallet’s extrinsics.